Refactoring: Extract a generic parent class
