+++ src/shogun2-core/shogun2-model/src/main/java/de/terrestris/shogun2/converter/UserGroupIdResolver.java
@@ -1,60 +1,30 @@
 package de.terrestris.shogun2.converter;
 
-import org.apache.log4j.Logger;
-
-import com.fasterxml.jackson.annotation.ObjectIdGenerator.IdKey;
-import com.fasterxml.jackson.annotation.ObjectIdResolver;
-import com.fasterxml.jackson.annotation.SimpleObjectIdResolver;
-
-import de.terrestris.shogun2.helper.IdHelper;
 import de.terrestris.shogun2.model.UserGroup;
 
 /**
- * 
- * An ID resolver for the {@link UserGroup} when deserializing only on the base
- * of ID values. Extends the default implementation.
  *
  * @author Nils Buehner
  *
  */
-public class UserGroupIdResolver extends SimpleObjectIdResolver {
+public class UserGroupIdResolver<E extends UserGroup> extends
+		PersistentObjectIdResolver<UserGroup> {
 
-	protected final Logger LOG = Logger.getLogger(getClass());
-
+	/**
+	 * Default Constructor
+	 */
 	public UserGroupIdResolver() {
+		super(UserGroup.class);
 	}
 
-	@Override
-	public void bindItem(IdKey id, Object ob) {
-		super.bindItem(id, ob);
+	/**
+	 * Constructor that has to be called by subclasses.
+	 *
+	 * @param entityClass
+	 */
+	@SuppressWarnings("unchecked")
+	protected UserGroupIdResolver(Class<E> entityClass) {
+		super((Class<UserGroup>) entityClass);
 	}
 
-	@Override
-	public Object resolveId(IdKey id) {
-
-		try {
-			if (id.key instanceof Integer) {
-				// return a "dummy" object, which only has the correct ID value
-				UserGroup userGroup = new UserGroup();
-				IdHelper.setIdOnPersistentObject(userGroup, (Integer) id.key);
-				return userGroup;
-			} else {
-				throw new Exception("ID is not of type Integer.");
-			}
-		} catch (Exception e) {
-			LOG.error("Could not resolve object by ID: " + e.getMessage());
-			return null;
-		}
-
-	}
-
-	@Override
-	public boolean canUseFor(ObjectIdResolver resolverType) {
-		return super.canUseFor(resolverType);
-	}
-
-	@Override
-	public ObjectIdResolver newForDeserialization(Object context) {
-		return new UserGroupIdResolver();
-	}
 }
